+++ library/src/utils/text.ts
@@ -1,7 +1,10 @@
+import type { Modifiers } from '../engine/types'
+
 export const isBoolString = (str: string) => str.trim() === 'true'
 
-export const lcFirst = (str: string) =>
-  str[0].toLowerCase() + str.slice(1)
+export const lcFirst = (str: string) => str[0].toLowerCase() + str.slice(1)
+
+export const ucFirst = (str: string) => str[0].toUpperCase() + str.slice(1)
 
 export const kebabize = (str: string) =>
   str.replace(
@@ -17,3 +20,18 @@ export const jsStrToObject = (raw: string) =>
 
 export const trimDollarSignPrefix = (str: string) =>
   str.startsWith('$') ? str.slice(1) : str
+
+export function modifyCasing(str: string, mods: Modifiers) {
+  const casing = mods.get('casing')
+  if (casing) {
+    str = casing.has('kebab')
+      ? kebabize(str)
+      : casing.has('snake')
+        ? kebabize(str).replace(/-/g, '_')
+        : casing.has('pascal')
+          ? ucFirst(str)
+          : str
+  }
+
+  return str
+}